Seperating devTools class
