+++ test-output/extentReport.html
@@ -35,7 +35,7 @@
 <a href="#"><span class="badge badge-primary">Test Report</span></a>
 </li>
 <li class="m-r-10">
-<a href="#"><span class="badge badge-primary">Monday, August 25, 2025, 03:34 PM (EEST)</span></a>
+<a href="#"><span class="badge badge-primary">Tuesday, August 26, 2025, 12:07 PM (EEST)</span></a>
 </li>
 </ul>
 </div>
@@ -87,13 +87,13 @@
 <div class="col-md-3">
 <div class="card"><div class="card-body">
 <p class="m-b-0">Started</p>
-<h3>Monday, August 25, 2025, 03:34 PM (EEST)</h3>
+<h3>Tuesday, August 26, 2025, 12:07 PM (EEST)</h3>
 </div></div>
 </div>
 <div class="col-md-3">
 <div class="card"><div class="card-body">
 <p class="m-b-0">Ended</p>
-<h3>Monday, August 25, 2025, 03:34 PM (EEST)</h3>
+<h3>Tuesday, August 26, 2025, 12:08 PM (EEST)</h3>
 </div></div>
 </div>
 <div class="col-md-3">
